wwan: remove all wwan related packages out of nss packages
These packages really shouldn't be in this repository as it's not the
core focus of the project. I am unable to test or maintain these since
it requires a device with a Quectel modem and cellular service, neither of
which I have access to.

For `Arcadyan AW1000` users you may be interested in trying:

https://github.com/immortalwrt/wwan-packages

All issues related to these packages will be closed and not addressed

Signed-off-by: Sean Khan <datapronix@protonmail.com>